requirements.txt
# requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
PyYAML>=6.0.0
tqdm>=4.66.0